pyproject.toml
[tool.poetry]
name = "ads111x"
version = "0.9.0"
description = "ADS111X series ADC tool for setup and data capture"
authors = ["Nikolay S. Vasil'ev <nickliamg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.18"
smbus2 = "^0.4.3"
colorama = "^0.4.6"
pytest = "^7.4.2"
click = "^8.1.7"
rpi-gpio = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
